+++ packages/terra-menu/CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 ### Breaking Changes
 * Consume latest major version bump of terra-popup
 
+5.10.0 - (May 7, 2019)
+------------------
 ### Changed
 * Added custom font-weight property for MenuItem component (defaults to 'inherit')